% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MitoTrace.R
\name{checkSequenceNames}
\alias{checkSequenceNames}
\title{MitoTrace - an R package for the investigation of mitochondrial heteroplasmies.}
\usage{
MitoTrace(bam_list = bams, ref_fasta = fasta_loc, name = "MT", max_depth = "", 
       min_base_quality = "", min_mapq = "", min_nucleotide_depth = "", min_minor_allele_depth = "")
}
\arguments{
\item{bams_list}{Vector of absolute path(s) pointing to BAM alignment file(s).}

\item{ref_fasta}{Absolute path to the mitochondrial reference genome in FASTA format.}

\item{name}{Name of mitochondrial genome as specified in the BAM files. Sequence names can be check with the checkSequenceNames() function.}

\item{tag_name}{The name of the tag corresponding the cellular barcode. Default = "CB". For droplet scRNAseq only.}

\item{min_read}{The minimum number of read counts to be considered a valid barcode (cell) in the analysis. Default = 1000. For droplet scRNAseq technologies only.}

\item{max_depth}{The maximum depth of reads considered at any position.}

\item{min_base_quality}{The minimum read base quality below which the base is ignored when summarizing pileup information.}

\item{min_mapq}{The minimum mapping quality below which the entire reads is ignored.}

\item{min_nucleotide_depth}{integer(1); minimum count of each nucleotide at a given position required for said nucleotide to appear in the result.}

\item{min_minor_allele_depth}{integer(1);  minimum count of all nucleotides other than the major allele at agiven position.}
}
\value{
Read counts matrix and coverage matrix.
}
\description{
The MitoTrace function calculates mitochondrial heteroplasmies in (single-cell) RNA sequencing data based on the reads pileups of the mitochondrial genome.
}
\details{
result <- MitoTrace(bam_list = bams, ref_fasta = fasta_loc, chr_name = "MT", max_depth = "1e6", min_base_quality=25, min_mapq=30, min_nucleotide_depth=0, min_minor_allele_depth=0)

See packageDescription("MitoTrace") for more details.
}
\note{
This package could not only apply for the analysis of single-cell data but also for bulk seuqencing data.
}
\references{
The current source code of MitoTrace is from https://github.com/lkmklsmn/MitoTrace.
}
\author{
Mingqiang WANG <Mingqiang.Wang@uth.tmc.edu>, Simon Lukas <lkmklsmn@gmail.com>
}
